v1.9.0 Beta 2
- Add furnishing timer
- Change timer page implement
- Add furnishing icons
- Add app action
- Update license
- Update Start Screen setting list
- Fix Radio dialog button size issue
- Fix Time Edit dialog hour string issue
- Fix Time Edit dialog hour digit limitations issue
- Add Timer Note Entry clear button
- Add Expedition Noti Note feature

+++ ResinTimer/ResinTimer/ResinTimer/Dialogs/TimeEditView.xaml.cs
@@ -33,7 +33,7 @@ namespace ResinTimer.Dialogs
 
             var ts = (noti.NotiTime > DateTime.Now) ? (noti.NotiTime - DateTime.Now) : TimeSpan.FromSeconds(0);
 
-            TimeHour.Text = ts.TotalHours.ToString();
+            TimeHour.Text = ((int)ts.TotalHours).ToString();
             TimeMinute.Text = ts.Minutes.ToString();
         }
